Renamed paramTree to ParameterTree.
